starting to dig out the depthmap

+++ lsd_slam_core/src/GlobalMapping/KeyFrameGraph.cpp
@@ -1,22 +1,22 @@
 /**
-* This file is part of LSD-SLAM.
-*
-* Copyright 2013 Jakob Engel <engelj at in dot tum dot de> (Technical University of Munich)
-* For more information see <http://vision.in.tum.de/lsdslam> 
-*
-* LSD-SLAM is free software: you can redistribute it and/or modify
-* it under the terms of the GNU General Public License as published by
-* the Free Software Foundation, either version 3 of the License, or
-* (at your option) any later version.
-*
-* LSD-SLAM is distributed in the hope that it will be useful,
-* but WITHOUT ANY WARRANTY; without even the implied warranty of
-* MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
-* GNU General Public License for more details.
-*
-* You should have received a copy of the GNU General Public License
-* along with LSD-SLAM. If not, see <http://www.gnu.org/licenses/>.
-*/
+ * This file is part of LSD-SLAM.
+ *
+ * Copyright 2013 Jakob Engel <engelj at in dot tum dot de> (Technical University of Munich)
+ * For more information see <http://vision.in.tum.de/lsdslam> 
+ *
+ * LSD-SLAM is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * LSD-SLAM is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with LSD-SLAM. If not, see <http://www.gnu.org/licenses/>.
+ */
 
 #include "GlobalMapping/KeyFrameGraph.h"
 #include "DataStructures/Frame.h"
@@ -37,7 +37,6 @@
 #include <g2o/types/sim3/sim3.h>
 #include "GlobalMapping/g2oTypeSim3Sophus.h"
 
-
 #include "IOWrapper/ImageDisplay.h"
 
 // for mkdir
@@ -52,42 +51,36 @@
 
 #include "util/globalFuncs.h"
 
-namespace lsd_slam
-{
+namespace lsd_slam {
 
-
-KFConstraintStruct::~KFConstraintStruct()
-{
-	if(edge != 0)
+KFConstraintStruct::~KFConstraintStruct() {
+	if (edge != 0)
 		delete edge;
 }
 
-KeyFrameGraph::KeyFrameGraph()
-: nextEdgeId(0)
-{
+KeyFrameGraph::KeyFrameGraph() :
+		nextEdgeId(0) {
 	typedef g2o::BlockSolver_7_3 BlockSolver;
 	typedef g2o::LinearSolverCSparse<BlockSolver::PoseMatrixType> LinearSolver;
 	//typedef g2o::LinearSolverPCG<BlockSolver::PoseMatrixType> LinearSolver;
 	LinearSolver* solver = new LinearSolver();
 	BlockSolver* blockSolver = new BlockSolver(solver);
-	g2o::OptimizationAlgorithmLevenberg* algorithm = new g2o::OptimizationAlgorithmLevenberg(blockSolver);
+	g2o::OptimizationAlgorithmLevenberg* algorithm =
+			new g2o::OptimizationAlgorithmLevenberg(blockSolver);
 	graph.setAlgorithm(algorithm);
-	
-    graph.setVerbose(false); // printOptimizationInfo
+
+	graph.setVerbose(false); // printOptimizationInfo
 	solver->setWriteDebug(true);
 	blockSolver->setWriteDebug(true);
 	algorithm->setWriteDebug(true);
 
-
-	totalPoints=0;
-	totalEdges=0;
-	totalVertices=0;
-
+	totalPoints = 0;
+	totalEdges = 0;
+	totalVertices = 0;
 
 }
 
-KeyFrameGraph::~KeyFrameGraph()
-{
+KeyFrameGraph::~KeyFrameGraph() {
 	// deletes edges
 	for (KFConstraintStruct* edge : newEdgeBuffer)
 		delete edge;	// deletes the g2oedge, which deletes the kernel.
@@ -98,49 +91,70 @@ KeyFrameGraph::~KeyFrameGraph()
 
 	// deletes pose structs (their memory is managed by graph)
 	// WARNING: at this point, all Frames have to be deleted, otherwise it night cause segfaults!
-	for(FramePoseStruct* p : allFramePoses)
+	for (FramePoseStruct* p : allFramePoses)
 		delete p;
 }
 
-
-void KeyFrameGraph::addFrame(Frame* frame)
-{
+void KeyFrameGraph::addFrame(Frame* frame) {
 
 	frame->pose->isRegisteredToGraph = true;
 	FramePoseStruct* pose = frame->pose;
-
 
 	allFramePosesMutex.lock();
 	allFramePoses.push_back(pose);
 	allFramePosesMutex.unlock();
 }
 
-void KeyFrameGraph::dumpMap(std::string folder)
-{
+void KeyFrameGraph::dumpMap(std::string folder) {
 	printf("DUMP MAP: dumping to %s\n", folder.c_str());
 
 	keyframesAllMutex.lock_shared();
-	char buf[100];
-	int succ = system(("rm -rf "+folder).c_str());
-	succ += system(("mkdir "+folder).c_str());
+	if (keyframesAll.empty()) {
+		keyframesAllMutex.unlock_shared();
+		std::cout << "DUMP MAP: nothing to dump. Aborting...\n";
+		return;
+	}
 
-	for(unsigned int i=0;i<keyframesAll.size();i++)
-	{
+	char buf[100];
+	int succ = system(("rm -rf " + folder).c_str());
+	succ += system(("mkdir " + folder).c_str());
+	std::cout << "DUMP MAP: succ = " << succ << std::endl;
+
+	for (unsigned int i = 0; i < keyframesAll.size(); i++) {
+		std::cout << "DUMP MAP: dumping keyframes\n";
 		snprintf(buf, 100, "%s/depth-%d.png", folder.c_str(), i);
 		cv::imwrite(buf, getDepthRainbowPlot(keyframesAll[i], 0));
 
+//		std::cout << "DUMP MAP: dumping depthmap for " << i << std::endl;
+//		for (unsigned int w = 0; w < keyframesAll[i]->width(0); w++) {
+//			for (unsigned int h = 0; h < keyframesAll[i]->height(0); h++) {
+//				std::cout
+//						<< (keyframesAll[i]->idepth(0))[w
+//								+ h * keyframesAll[i]->width(0)] << " ";
+//			}
+//			std::cout << std::endl;
+//		}
+
 		snprintf(buf, 100, "%s/frame-%d.png", folder.c_str(), i);
-		cv::imwrite(buf, cv::Mat(keyframesAll[i]->height(), keyframesAll[i]->width(),CV_32F,keyframesAll[i]->image()));
+		cv::imwrite(buf,
+				cv::Mat(keyframesAll[i]->height(), keyframesAll[i]->width(),
+						CV_32F, keyframesAll[i]->image()));
 
 		snprintf(buf, 100, "%s/var-%d.png", folder.c_str(), i);
-		cv::imwrite(buf, getVarRedGreenPlot(keyframesAll[i]->idepthVar(),keyframesAll[i]->image(),keyframesAll[i]->width(),keyframesAll[i]->height()));
+		cv::imwrite(buf,
+				getVarRedGreenPlot(keyframesAll[i]->idepthVar(),
+						keyframesAll[i]->image(), keyframesAll[i]->width(),
+						keyframesAll[i]->height()));
 	}
 
+	int i = keyframesAll.size() - 1;
+	Util::displayImage("VAR PREVIEW",
+			getVarRedGreenPlot(keyframesAll[i]->idepthVar(),
+					keyframesAll[i]->image(), keyframesAll[i]->width(),
+					keyframesAll[i]->height()));
 
-	int i = keyframesAll.size()-1;
-	Util::displayImage("VAR PREVIEW", getVarRedGreenPlot(keyframesAll[i]->idepthVar(),keyframesAll[i]->image(),keyframesAll[i]->width(),keyframesAll[i]->height()));
-
-	printf("DUMP MAP (succ %d): dumped %d depthmaps\n", succ,  (int)keyframesAll.size());
+	printf("DUMP MAP (succ %d): dumped %d depthmaps\n", succ,
+			(int) keyframesAll.size());
 
 	Eigen::MatrixXf res, resD, resP, huber, usage, consistency, distance, error;
 	Eigen::VectorXf meanRootInformation, usedPixels;
@@ -164,76 +178,79 @@ void KeyFrameGraph::dumpMap(std::string folder)
 	meanRootInformation.setZero();
 	usedPixels.setZero();
 
-	for(unsigned int i=0;i<keyframesAll.size();i++)
-	{
+	for (unsigned int i = 0; i < keyframesAll.size(); i++) {
 		meanRootInformation[i] = keyframesAll[i]->meanInformation;
-		usedPixels[i] = keyframesAll[i]->numPoints / (float)keyframesAll[i]->numMappablePixels;
+		usedPixels[i] = keyframesAll[i]->numPoints
+				/ (float) keyframesAll[i]->numMappablePixels;
 	}
 
-
 	edgesListsMutex.lock_shared();
-	for(unsigned int i=0;i<edgesAll.size();i++)
-	{
+	for (unsigned int i = 0; i < edgesAll.size(); i++) {
 		KFConstraintStruct* e = edgesAll[i];
 
-		res(e->firstFrame->idxInKeyframes, e->secondFrame->idxInKeyframes) = e->meanResidual;
-		resD(e->firstFrame->idxInKeyframes, e->secondFrame->idxInKeyframes) = e->meanResidualD;
-		resP(e->firstFrame->idxInKeyframes, e->secondFrame->idxInKeyframes) = e->meanResidualP;
-		usage(e->firstFrame->idxInKeyframes, e->secondFrame->idxInKeyframes) = e->usage;
-		consistency(e->firstFrame->idxInKeyframes, e->secondFrame->idxInKeyframes) = e->reciprocalConsistency;
-		distance(e->firstFrame->idxInKeyframes, e->secondFrame->idxInKeyframes) = e->secondToFirst.translation().norm();
-		error(e->firstFrame->idxInKeyframes, e->secondFrame->idxInKeyframes) = e->edge->chi2();
+		res(e->firstFrame->idxInKeyframes, e->secondFrame->idxInKeyframes) =
+				e->meanResidual;
+		resD(e->firstFrame->idxInKeyframes, e->secondFrame->idxInKeyframes) =
+				e->meanResidualD;
+		resP(e->firstFrame->idxInKeyframes, e->secondFrame->idxInKeyframes) =
+				e->meanResidualP;
+		usage(e->firstFrame->idxInKeyframes, e->secondFrame->idxInKeyframes) =
+				e->usage;
+		consistency(e->firstFrame->idxInKeyframes,
+				e->secondFrame->idxInKeyframes) = e->reciprocalConsistency;
+		distance(e->firstFrame->idxInKeyframes, e->secondFrame->idxInKeyframes) =
+				e->secondToFirst.translation().norm();
+		error(e->firstFrame->idxInKeyframes, e->secondFrame->idxInKeyframes) =
+				e->edge->chi2();
 	}
 	edgesListsMutex.unlock_shared();
 	keyframesAllMutex.unlock_shared();
 
-
 	std::ofstream fle;
 
-	fle.open(folder+"/residual.txt");
+	fle.open(folder + "/residual.txt");
 	fle << res;
 	fle.close();
+	std::cout << "DUMP MAP: res: \n" << res << std::endl;
 
-	fle.open(folder+"/residualD.txt");
+	fle.open(folder + "/residualD.txt");
 	fle << resD;
 	fle.close();
 
-	fle.open(folder+"/residualP.txt");
+	fle.open(folder + "/residualP.txt");
 	fle << resP;
 	fle.close();
 
-	fle.open(folder+"/usage.txt");
+	fle.open(folder + "/usage.txt");
 	fle << usage;
 	fle.close();
 
-	fle.open(folder+"/consistency.txt");
+	fle.open(folder + "/consistency.txt");
 	fle << consistency;
 	fle.close();
 
-	fle.open(folder+"/distance.txt");
+	fle.open(folder + "/distance.txt");
 	fle << distance;
 	fle.close();
+	std::cout << "DUMP MAP: dist: \n" << distance << std::endl;
 
-	fle.open(folder+"/error.txt");
+	fle.open(folder + "/error.txt");
 	fle << error;
 	fle.close();
 
-	fle.open(folder+"/meanRootInformation.txt");
+	fle.open(folder + "/meanRootInformation.txt");
 	fle << meanRootInformation;
 	fle.close();
 
-	fle.open(folder+"/usedPixels.txt");
+	fle.open(folder + "/usedPixels.txt");
 	fle << usedPixels;
 	fle.close();
 
-	printf("DUMP MAP: dumped %d edges\n", (int)edgesAll.size());
+	printf("DUMP MAP: dumped %d edges\n", (int) edgesAll.size());
 }
 
-
-
-void KeyFrameGraph::addKeyFrame(Frame* frame)
-{
-	if(frame->pose->graphVertex != nullptr)
+void KeyFrameGraph::addKeyFrame(Frame* frame) {
+	if (frame->pose->graphVertex != nullptr)
 		return;
 
 	// Insert vertex into g2o graph
@@ -242,7 +259,7 @@ void KeyFrameGraph::addKeyFrame(Frame* frame)
 
 	Sophus::Sim3d camToWorld_estimate = frame->getScaledCamToWorld();
 
-	if(!frame->hasTrackingParent())
+	if (!frame->hasTrackingParent())
 		vertex->setFixed(true);
 
 	vertex->setEstimate(camToWorld_estimate);
@@ -254,11 +271,10 @@ void KeyFrameGraph::addKeyFrame(Frame* frame)
 
 }
 
-void KeyFrameGraph::insertConstraint(KFConstraintStruct* constraint)
-{
+void KeyFrameGraph::insertConstraint(KFConstraintStruct* constraint) {
 	EdgeSim3* edge = new EdgeSim3();
 	edge->setId(nextEdgeId);
-	++ nextEdgeId;
+	++nextEdgeId;
 
 	totalEdges++;
 
@@ -275,16 +291,12 @@ void KeyFrameGraph::insertConstraint(KFConstraintStruct* constraint)
 	constraint->edge = edge;
 	newEdgeBuffer.push_back(constraint);
 
-
 	constraint->firstFrame->neighbors.insert(constraint->secondFrame);
 	constraint->secondFrame->neighbors.insert(constraint->firstFrame);
 
-	for(int i=0;i<totalVertices;i++)
-	{
+	for (int i = 0; i < totalVertices; i++) {
 		//shortestDistancesMap
 	}
-
-
 
 	edgesListsMutex.lock();
 	constraint->idxInAllEdges = edgesAll.size();
@@ -292,27 +304,27 @@ void KeyFrameGraph::insertConstraint(KFConstraintStruct* constraint)
 	edgesListsMutex.unlock();
 }
 
-
-bool KeyFrameGraph::addElementsFromBuffer()
-{
+bool KeyFrameGraph::addElementsFromBuffer() {
 	bool added = false;
 
 	keyframesForRetrackMutex.lock();
-	for (auto newKF : newKeyframesBuffer)
-	{
+	keyframesAllMutex.lock_shared();
+	for (auto newKF : newKeyframesBuffer) {
 		graph.addVertex(newKF->pose->graphVertex);
 		assert(!newKF->pose->isInGraph);
 		newKF->pose->isInGraph = true;
 
 		keyframesForRetrack.push_back(newKF);
+		std::cout << "ADDING elements from KF buffer\n";
+		keyframesAll.push_back(newKF);
 
 		added = true;
 	}
+	keyframesAllMutex.unlock_shared();
 	keyframesForRetrackMutex.unlock();
 
 	newKeyframesBuffer.clear();
-	for (auto edge : newEdgeBuffer)
-	{
+	for (auto edge : newEdgeBuffer) {
 		graph.addEdge(edge->edge);
 		added = true;
 	}
@@ -321,49 +333,43 @@ bool KeyFrameGraph::addElementsFromBuffer()
 	return added;
 }
 
-int KeyFrameGraph::optimize(int num_iterations)
-{
+int KeyFrameGraph::optimize(int num_iterations) {
 	// Abort if graph is empty, g2o shows an error otherwise
 	if (graph.edges().size() == 0)
 		return 0;
-	
+
 	graph.setVerbose(false); // printOptimizationInfo
 	graph.initializeOptimization();
-	
 
 	return graph.optimize(num_iterations, false);
 
 }
 
-
-
-void KeyFrameGraph::calculateGraphDistancesToFrame(Frame* startFrame, std::unordered_map< Frame*, int >* distanceMap)
-{
+void KeyFrameGraph::calculateGraphDistancesToFrame(Frame* startFrame,
+		std::unordered_map<Frame*, int>* distanceMap) {
 	distanceMap->insert(std::make_pair(startFrame, 0));
-	
-	std::multimap< int, Frame* > priorityQueue;
+
+	std::multimap<int, Frame*> priorityQueue;
 	priorityQueue.insert(std::make_pair(0, startFrame));
-	while (! priorityQueue.empty())
-	{
+	while (!priorityQueue.empty()) {
 		auto it = priorityQueue.begin();
 		int length = it->first;
 		Frame* frame = it->second;
 		priorityQueue.erase(it);
-		
+
 		auto mapEntry = distanceMap->find(frame);
-		
-		if (mapEntry != distanceMap->end() && length > mapEntry->second)
-		{
+
+		if (mapEntry != distanceMap->end() && length > mapEntry->second) {
 			continue;
 		}
-		
-		for (Frame* neighbor : frame->neighbors)
-		{
+
+		for (Frame* neighbor : frame->neighbors) {
 			auto neighborMapEntry = distanceMap->find(neighbor);
-			
-			if (neighborMapEntry != distanceMap->end() && length + 1 >= neighborMapEntry->second)
+
+			if (neighborMapEntry != distanceMap->end()
+					&& length + 1 >= neighborMapEntry->second)
 				continue;
-			
+
 			if (neighborMapEntry != distanceMap->end())
 				neighborMapEntry->second = length + 1;
 			else
